# Conflicts:
#   app/src/main/java/com/android/filemaster/data/repository/FileRepository.kt
#   app/src/main/java/com/android/filemaster/data/viewmodel/FileViewModel.kt
#   app/src/main/java/com/android/filemaster/module/AppExt.kt
#   app/src/main/java/com/android/filemaster/ui/customview/CircularProgressBar.kt
#   app/src/main/java/com/android/filemaster/ui/home/HomeFragment.kt
#   app/src/main/java/com/android/filemaster/utils/DataFake.kt
#   app/src/main/java/com/android/filemaster/utils/FileManager.kt
#   app/src/main/res/layout/fragment_home.xml
#   app/src/main/res/layout/item_strorage.xml

+++ app/src/main/java/com/android/filemaster/ui/customview/CircularProgressBar.kt
@@ -83,14 +83,14 @@ class CircularProgressBar(context: Context?, attrs: AttributeSet?, defStyleAttr:
 
     }
 
-    private fun calcSweepAngleFromProgress(maxProgress: Long, progress: Long): Float {
+    private fun calcSweepAngleFromProgress(maxProgress: Int, progress: Int): Float {
         return mMaxSweepAngle / maxProgress * progress
     }
 
 
-    fun setProgress(maxProgress: Long, progress: Long) {
+    fun setProgress(maxProgress: Int, progress: Int) {
         val animator =
-                ValueAnimator.ofFloat(mSweepAngleBackground, calcSweepAngleFromProgress(maxProgress, progress))
+            ValueAnimator.ofFloat(mSweepAngleBackground, calcSweepAngleFromProgress(maxProgress, progress))
         animator.interpolator = DecelerateInterpolator()
         animator.duration = mAnimationDuration.toLong()
         animator.addUpdateListener { valueAnimator ->
